requirements.txt
openai==0.27.5
python-dotenv==1.0.0
pandas
wordcloud==1.9.2
matplotlib
plotly==5.15.0
nltk==3.8.1
Pillow==9.5.0
streamlit-ace==0.1.1
extra-streamlit-components==0.1.56
streamlit-image-select==0.6.0
sewar==0.4.6
shapely==2.0.1

